:zap: viewcampupdate screen + styles. created a new file in constants
